fix(validate): Clippy lints and safety comments in validator

+++ crates/expanse/src/map.rs
@@ -438,7 +438,7 @@ impl ExpanseMap {
             Root::Empty => {}
             Root::Leaf { pop, .. } => {
                 stats.depth_histogram[0] = 1;
-                stats.leaf_pop_histogram[*pop as usize] = 1;
+                stats.leaf_pop_histogram[*pop] = 1;
                 stats.node_counts.leaf_linear = 1;
             }
             Root::Tree { top, .. } => {
